objc2_encode

Enum EncodingBox

#[non_exhaustive]
pub enum EncodingBox {

    Char,
    Short,
    Int,
    Long,
    LongLong,
    UChar,
    UShort,
    UInt,
    ULong,
    ULongLong,
    Float,
    Double,
    LongDouble,
    FloatComplex,
    DoubleComplex,
    LongDoubleComplex,
    Bool,
    Void,
    String,
    Object,
    Block,
    Class,
    Sel,
    Unknown,
    BitField(u8, Option<Box<(u64, Self)>>),
    Pointer(Box<Self>),
    Atomic(Box<Self>),
    Array(u64, Box<Self>),
    Struct(String, Vec<Self>),
    Union(String, Vec<Self>),
    None,

}

The boxed version of Encoding.

This has exactly the same items as Encoding, the only difference is in where the contents of the more complex encodings like Struct are stored.

In Encoding, the data is stored in static memory, while in EncodingBox it is stored on the heap. The former allows storing in constants (which is required by the objc2::encode::Encode and objc2::encode::RefEncode traits), while the latter allows dynamic creation, such as in the case of parsing encodings.

This should be considered a temporary restriction. Encoding and EncodingBox will become equivalent once heap allocation in constants is possible.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Char

Same as Encoding::Char.

Short

Same as Encoding::Short.

Int

Same as Encoding::Int.

Long

Same as Encoding::Long.

LongLong

Same as Encoding::LongLong.

UChar

Same as Encoding::UChar.

UShort

Same as Encoding::UShort.

UInt

Same as Encoding::UInt.

ULong

Same as Encoding::ULong.

ULongLong

Same as Encoding::ULongLong.

Float

Same as Encoding::Float.

Double

Same as Encoding::Double.

LongDouble

Same as Encoding::LongDouble.

FloatComplex

Same as Encoding::FloatComplex.

DoubleComplex

Same as Encoding::DoubleComplex.

LongDoubleComplex

Same as Encoding::LongDoubleComplex.

Bool

Same as Encoding::Bool.

Void

Same as Encoding::Void.

String

Same as Encoding::String.

Object

Same as Encoding::Object.

Block

Same as Encoding::Block.

Class

Same as Encoding::Class.

Sel

Same as Encoding::Sel.

Unknown

Same as Encoding::Unknown.

BitField(u8, Option<Box<(u64, Self)>>)

Same as Encoding::BitField.

Pointer(Box<Self>)

Same as Encoding::Pointer.

Atomic(Box<Self>)

Same as Encoding::Atomic.

Array(u64, Box<Self>)

Same as Encoding::Array.

Struct(String, Vec<Self>)

Same as Encoding::Struct.

Union(String, Vec<Self>)

Same as Encoding::Union.

None

Same as Encoding::None.

Implementations

impl EncodingBox

pub const C_LONG: Self

Same as Encoding::C_LONG.

pub const C_ULONG: Self

Same as Encoding::C_ULONG.

pub fn from_start_of_str(s: &mut &str) -> Result<Self, ParseError>

Parse and consume an encoding from the start of a string.

This is can be used to parse concatenated encodings, such as those returned by method_getTypeEncoding.

from_str is simpler, use that instead if you can.

Errors

Returns an error if the string was an ill-formatted encoding string.