Trait objc2_encode::EncodeArguments

pub unsafe trait EncodeArguments: Sealed {
    const ENCODINGS: &'static [Encoding];
}

Types that represent an ordered group of function arguments, where each argument has an Objective-C type-encoding, or can be converted from one.

This is implemented for tuples of up to 12 arguments, where each argument implements EncodeConvert. It is primarily used to make generic code a bit easier.

Note that tuples themselves don’t implement Encode directly, because they’re not FFI-safe!

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!

Safety

You cannot rely on this trait for ensuring that the arguments all are Encode, since they may only implement EncodeConvert. Use your own trait and add Encode bounds on that.

Required Associated Constants

const ENCODINGS: &'static [Encoding]

The encodings for the arguments.